Dissolved AppMotor.CliApp.Hosting and moved its class into AppMotor.CliApp
This makes the project structure much simpler
