+++ AppMotor.sln
@@ -38,8 +38,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AppMotor.CliApp", "src\AppM
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AppMotor.CliApp.Tests", "tests\AppMotor.CliApp.Tests\AppMotor.CliApp.Tests.csproj", "{F4EAF0C1-51E2-40BD-8A54-627D8D958428}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AppMotor.CliApp.Hosting", "src\AppMotor.CliApp.Hosting\AppMotor.CliApp.Hosting.csproj", "{4DA273A3-4340-4713-BF15-EFE5685CDA23}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AppMotor.CliApp.Samples.AppWithVerbs", "samples\AppMotor.CliApp.Samples.AppWithVerbs\AppMotor.CliApp.Samples.AppWithVerbs.csproj", "{40B23696-445C-42B2-85AA-5ED5A3EF428B}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AppMotor.CliApp.Samples.AppWithParams", "samples\AppMotor.CliApp.Samples.AppWithParams\AppMotor.CliApp.Samples.AppWithParams.csproj", "{33259BEB-A97E-4869-B1F2-9B668C2A00B9}"
@@ -82,10 +80,6 @@ Global
 		{F4EAF0C1-51E2-40BD-8A54-627D8D958428}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F4EAF0C1-51E2-40BD-8A54-627D8D958428}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F4EAF0C1-51E2-40BD-8A54-627D8D958428}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4DA273A3-4340-4713-BF15-EFE5685CDA23}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4DA273A3-4340-4713-BF15-EFE5685CDA23}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4DA273A3-4340-4713-BF15-EFE5685CDA23}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4DA273A3-4340-4713-BF15-EFE5685CDA23}.Release|Any CPU.Build.0 = Release|Any CPU
 		{40B23696-445C-42B2-85AA-5ED5A3EF428B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{40B23696-445C-42B2-85AA-5ED5A3EF428B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{40B23696-445C-42B2-85AA-5ED5A3EF428B}.Release|Any CPU.ActiveCfg = Release|Any CPU
