+++ README.md
@@ -4,12 +4,11 @@ This repository contains the code for the AppMotor packages. AppMotor provide va
 
 ## Packages
 
-| Package                   | Version                                                                                                                           | Description
-| ------------------------- | --------------------------------------------------------------------------------------------------------------------------------- | -----------
-| `AppMotor.Core`           | [![NuGet](https://img.shields.io/nuget/v/AppMotor.Core.svg)](https://www.nuget.org/packages/AppMotor.Core/)                       | Core APIs
-| `AppMotor.CliApp`         | [![NuGet](https://img.shields.io/nuget/v/AppMotor.CliApp.svg)](https://www.nuget.org/packages/AppMotor.CliApp/)                   | APIs for command line applications; see [README.md](src/AppMotor.CliApp/README.md) for details
-| `AppMotor.CliApp.Hosting` | [![NuGet](https://img.shields.io/nuget/v/AppMotor.CliApp.Hosting.svg)](https://www.nuget.org/packages/AppMotor.CliApp.Hosting/)   | APIs for combining `AppMotor.CliApp` and .NET's [generic host](https://docs.microsoft.com/en-us/dotnet/core/extensions/generic-host)
-| `AppMotor.HttpServer`     | [![NuGet](https://img.shields.io/nuget/v/AppMotor.HttpServer.svg)](https://www.nuget.org/packages/AppMotor.HttpServer/)            | APIs for running an HTTP(S) server; see [README.md](src/AppMotor.HttpServer/README.md) for details
+| Package                   | Version                                                                                                                 | Description
+| ------------------------- | ----------------------------------------------------------------------------------------------------------------------- | -----------
+| `AppMotor.Core`           | [![NuGet](https://img.shields.io/nuget/v/AppMotor.Core.svg)](https://www.nuget.org/packages/AppMotor.Core/)             | Core APIs
+| `AppMotor.CliApp`         | [![NuGet](https://img.shields.io/nuget/v/AppMotor.CliApp.svg)](https://www.nuget.org/packages/AppMotor.CliApp/)         | APIs for command line applications; see [README.md](src/AppMotor.CliApp/README.md) for details
+| `AppMotor.HttpServer`     | [![NuGet](https://img.shields.io/nuget/v/AppMotor.HttpServer.svg)](https://www.nuget.org/packages/AppMotor.HttpServer/) | APIs for running an HTTP(S) server; see [README.md](src/AppMotor.HttpServer/README.md) for details
 
 ## Note on Breaking Changes
