--- src/AppMotor.CliApp.Hosting/AppMotor.CliApp.Hosting.NuGet.props
@@ -1,19 +1,0 @@
-<!--
-  Contains NuGet information for this project.
-
-  NOTE: This file is automatically imported by MSBuild.
--->
-<Project>
-
-  <!--
-    See 'Project.NuGetPackage.props' for more details.
-  -->
-  <PropertyGroup>
-    <Authors>Sebastian Krysmanski</Authors>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <CommonlyUsedTypes Include="AppMotor.CliApp.Hosting.CliCommandWithGenericHost" />
-  </ItemGroup>
-
-</Project>
