--- src/AppMotor.CliApp.Hosting/AppMotor.CliApp.Hosting.csproj
@@ -1,24 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="$(ProjectCommonsDirectory)Project.Code.props" />
-  <Import Project="$(ProjectCommonsDirectory)Project.NuGetPackage.props" />
-  <Import Project="../Common.props" />
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-
-    <!-- See also: https://docs.microsoft.com/en-us/nuget/create-packages/prerelease-packages -->
-    <Version>$(AppMotorCliAppVersion)</Version>
-
-    <Description>APIs for combining "AppMotor.CliApp" with "Microsoft.Extensions.Hosting".</Description>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="5.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\AppMotor.CliApp\AppMotor.CliApp.csproj" />
-  </ItemGroup>
-
-</Project>
