+++ src/AppMotor.HttpServer/HttpServerCommandBase.cs
@@ -4,7 +4,6 @@ using System.Net;
 using System.Security.Cryptography.X509Certificates;
 
 using AppMotor.CliApp.CommandLine;
-using AppMotor.CliApp.Hosting;
 using AppMotor.Core.Certificates;
 using AppMotor.Core.Exceptions;
 using AppMotor.Core.Net;
